Date format

+++ Clima/Clima/Model/ForecastData.swift
@@ -13,7 +13,7 @@ struct ForecastData: Codable {
 }
 
 struct List: Codable {
-    let dt: Int
+    let dt: Double
     let main: ForecastMain
     let weather: [ForecastWeather]
 }
